# annotation/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BertopicLabels": {
        "description": "Distinct BERTopic topics + subtopics (with parent topic) and counts.",
        "properties": {
          "subtopics": {
            "items": {
              "$ref": "#/components/schemas/BertopicSubtopicCount"
            },
            "title": "Subtopics",
            "type": "array"
          },
          "topics": {
            "items": {
              "$ref": "#/components/schemas/BertopicTopicCount"
            },
            "title": "Topics",
            "type": "array"
          }
        },
        "title": "BertopicLabels",
        "type": "object"
      },
...
      "BoundaryRequest": {
        "description": "Replace a conversation's gold spans (split/merge).",
        "properties": {
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/BoundarySpan"
            },
            "title": "Segments",
            "type": "array"
          }
        },
        "required": [
          "segments"
        ],
        "title": "BoundaryRequest",
        "type": "object"
      },
...
      "ConversationPage": {
        "description": "A paginated page of conversation summaries for the review queue.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConversationSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "page",
          "page_size"
        ],
        "title": "ConversationPage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Labelers": {
        "description": "Distinct worklist labelers for a dataset, sorted (may be empty).",
        "properties": {
          "labelers": {
            "items": {
              "type": "string"
            },
            "title": "Labelers",
            "type": "array"
          }
        },
        "title": "Labelers",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "eb1 Annotation Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/datasets": {
      "get": {
        "description": "List registered dataset names.",
        "operationId": "list_datasets_api_datasets_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response List Datasets Api Datasets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Datasets"
      }
    },
    "/api/datasets/{dataset}/bertopic-labels": {
      "get": {
        "description": "Return distinct BERTopic topics + subtopics with per-conversation counts.",
        "operationId": "get_bertopic_labels_api_datasets__dataset__bertopic_labels_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset",
            "required": true,
            "schema": {
              "title": "Dataset",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BertopicLabels"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Bertopic Labels"
      }
    },
    "/api/datasets/{dataset}/conversations": {
      "get": {
        "description": "Return a PAGINATED, searchable page of conversation summaries.\n\n``q`` matches conversation ext_id OR message content. ``status``/``topic``\nfilter on the effective segmentation. ``labeler`` restricts the page to that\nlabeler's worklist assignments. ``bertopic_topic``/``bertopic_subtopic``\nrestrict to conversations with a gold segment carrying that BERTopic label.\n``mismatch`` restricts to conversations with a reviewed gold segment whose\nhuman label disagrees with its BERTopic source label. Shape:\n``{items,total,page,page_size}``.",
        "operationId": "list_conversations_api_datasets__dataset__conversations_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset",
            "required": true,
            "schema": {
              "title": "Dataset",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          },
          {
            "in": "query",
            "name": "labeler",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Labeler"
            }
          },
          {
            "in": "query",
            "name": "bertopic_topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bertopic Topic"
            }
          },
          {
            "in": "query",
            "name": "bertopic_subtopic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bertopic Subtopic"
            }
          },
          {
            "in": "query",
            "name": "mismatch",
            "required": false,
            "schema": {
              "default": false,
              "title": "Mismatch",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Conversations"
      }
    },
    "/api/datasets/{dataset}/conversations/{conversation}": {
      "get": {
        "description": "Return a conversation's messages plus its effective + gold segments.",
        "operationId": "get_conversation_api_datasets__dataset__conversations__conversation__get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset",
            "required": true,
            "schema": {
              "title": "Dataset",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "conversation",
            "required": true,
            "schema": {
              "title": "Conversation",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Conversation"
      }
    },
    "/api/datasets/{dataset}/conversations/{conversation}/boundaries": {
      "post": {
        "description": "REPLACE all gold for a conversation with the posted spans (split/merge).\n\nEach new span inherits topic/subtopic/sentiment from the overlapping\npredicted segment when the client leaves them unset.",
        "operationId": "replace_boundaries_api_datasets__dataset__conversations__conversation__boundaries_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset",
            "required": true,
            "schema": {
              "title": "Dataset",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "conversation",
            "required": true,
            "schema": {
              "title": "Conversation",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoundaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoundaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Boundaries"
      }
    },
    "/api/datasets/{dataset}/labelers": {
      "get": {
        "description": "Return the distinct worklist labelers for the dataset, sorted (may be empty).",
        "operationId": "get_labelers_api_datasets__dataset__labelers_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset",
            "required": true,
            "schema": {
              "title": "Dataset",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Labelers"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Labelers"
      }
    },
    "/api/datasets/{dataset}/segments": {
      "get": {
        "description": "Return the predicted-segment review queue with gold review state.\n\nFilters (optional): ``status`` (``reviewed``/``unreviewed``), ``topic`` exact\nmatch, ``max_confidence`` (keep segments at or below that confidence).",
        "operationId": "list_segments_api_datasets__dataset__segments_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset",
            "required": true,
            "schema": {
              "title": "Dataset",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          },
          {
            "in": "query",
            "name": "max_confidence",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Confidence"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SegmentSummary"
                  },
                  "title": "Response List Segments Api Datasets  Dataset  Segments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Segments"
      }
    },
...
  }
}